TDS/requirements.txt
PyAudio==0.2.12
SpeechRecognition==3.8.1
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.5
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
praat-parselmouth==0.4.2
PyAudio==0.2.12
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
Wave==0.0.2
wavio==0.0.7
